+++ ui/src/kapacitor/containers/KapacitorRulesPage.tsx
@@ -109,6 +109,7 @@ export class KapacitorRulesPage extends PureComponent<Props, State> {
       <KapacitorRules
         rules={rules}
         source={source}
+        kapacitor={kapacitor}
         onDelete={this.handleDeleteRule}
         onChangeRuleStatus={this.handleRuleStatus}
       />